% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{Covid19IncidenceRasInhibitors}
\alias{Covid19IncidenceRasInhibitors}
\title{Covid19IncidenceRasInhibitors}
\description{
An OHDSI R package to compare the relative risk of COVID-19 susceptibility for RAS inhibitor users
}
